Move ConfigureGoogleAccount, SyncGoogleVCs
